#RR change ads CRUD and apply full api test

+++ src/main/java/auto/freitagsmarkt/service/impl/AdsServiceImpl.java
@@ -1,11 +1,12 @@
 package auto.freitagsmarkt.service.impl;
 
-
 import auto.freitagsmarkt.domain.Ads;
 import auto.freitagsmarkt.dto.AdsDTO;
 import auto.freitagsmarkt.mapper.AdsMapper;
 import auto.freitagsmarkt.repository.AdsRepository;
 import auto.freitagsmarkt.service.AdsService;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
@@ -14,7 +15,7 @@ import java.util.*;
 public class AdsServiceImpl implements AdsService {
     private AdsRepository adsRepository;
     private AdsMapper adsMapper;
-
+    private final static Logger log = LoggerFactory.getLogger(AdsServiceImpl.class);
     public AdsServiceImpl(AdsRepository adsRepository, AdsMapper adsMapper) {
         this.adsRepository = adsRepository;
         this.adsMapper = adsMapper;
@@ -31,7 +32,7 @@ public class AdsServiceImpl implements AdsService {
 
     @Override
     public AdsDTO createNewAd(AdsDTO adsDTO) {
-        return Optional.of(adsDTO)
+        return Optional.ofNullable(adsDTO)
                 .map(adsMapper::toAd)
                 .map(adsRepository::save)
                 .map(adsMapper::toAdDTO)
@@ -54,7 +55,11 @@ public class AdsServiceImpl implements AdsService {
     }
 
     @Override
-    public void removeAdById(Long adId) {
-        adsRepository.deleteById(adId);
+    public Boolean removeAdById(Long adId) {
+        if(adsRepository.existsById(adId)){
+            adsRepository.deleteById(adId);
+            return Boolean.TRUE;
+        }
+        return Boolean.FALSE;
     }
 }
